services/transcription/requirements.txt
flask
python-dotenv
pydub
openai

# system dependency: ffmpeg (install via brew/apt)